+++ app/Http/Controllers/Api/ScoreController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ScoreResource;
 use App\Models\Score;
+use Illuminate\Support\Facades\Gate;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class ScoreController extends Controller
@@ -14,6 +15,7 @@ class ScoreController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Score::class);
         $scores = QueryBuilder::for(Score::class)
             ->allowedIncludes('score_able')
             ->get();
@@ -30,6 +32,7 @@ class ScoreController extends Controller
      */
     public function show(Score $score)
     {
+        Gate::authorize('view', Score::class);
         return $this->successResponse(
             ScoreResource::make($score->load('score_able')),
             'Successfully find score'
@@ -41,6 +44,7 @@ class ScoreController extends Controller
      */
     public function destroy(Score $score)
     {
+        Gate::authorize('delete', Score::class);
         $score->delete();
 
         return $this->successResponse(message: 'Successfully deleted score');
